+++ src/main/java/com/fitwsarah/fitwsarah/authenticationsubdomain/presentationlayer/Auth0Controller.java
@@ -1,8 +1,6 @@
 package com.fitwsarah.fitwsarah.authenticationsubdomain.presentationlayer;
 
-import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
-import org.springframework.security.config.annotation.web.builders.HttpSecurity;
 import org.springframework.web.bind.annotation.CrossOrigin;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
